refactor: rename temp stabletoken.1 component + delete old component
